Remove init.vim and rename settings
