--- nvim/init.vim
@@ -1,4 +1,0 @@
-set runtimepath^=~/.vim runtimepath+=~/.vim/after runtimepath+=~/.config/nvim/lua
-let &packpath=&runtimepath
-source ~/.vimrc
-lua require('plugins')
