+++ nvim/lua/akan72/init.lua
@@ -1,3 +1,3 @@
 require("akan72.plugins")
 require("akan72.remaps")
-require("akan72.set")
+require("akan72.settings")
