+++ nvim/lua/akan72/settings.lua
